Added function to download the cross toolchain
This commit introduces the new function `ensure_toolchain_installed()`
which dowloads the cross compiler toolchain for the Roborio.

Unfortunatly it is not possible in CMake to build for multiple
platforms in the same build, so there will need to be something done
about that. Maybe having seperate subdirectories for each build with
their own CMakeLists.txt in them?