CMake can compile for both the native (maybe) and target platforms
Currently the frcUserProgram can compile with **my** native platform
(Linux x86_64) and has not been tested with any other platform as of
yet. The cross compilation has also only been tested on this platform
currently. As long as you use Linux x86_64 you are guaranteed for it
to work.

In order to compile using CMake you have to link to the
correct libraries. The file deps/allwpilib.json is pretty much the
same as json files for vender dependancies, so it can be used in the
`ensure_vendors_installed()` function as well, and that will download
all the WPILib libraries that you need to link to. WPILib also depends
on opencv, which there is also a json file deps/opencv-cpp.json to use
as well.

When using the function `ensure_vendors_installed()` it will create
a directory in ${CMAKE_SOURCE_DIR} (the root of your project) called
.vendor and thats where it will store all of its information. In order
to properly link to the libraries it creates inside that directory
cmake/Modules/Find<VenderDependancy>.cmake which can be found with cmake
using the `find_package()` function which will include targets for
everything in that vendor dependancy json. Before you can do this you
have to set the CMake variable `CMAKE_MODULE_PATH` to the cmake/Modules
path for CMake to be able to find them during the `find_package()` call.

Unfortunatly CMake does not allow for different compilers and toolchains
for one build, so we have to split up the native and target build into
two seperate projects.