Fixed cross and native compilation for MacOS
The `find_library()` calls in the configured files
were not taking the version numbers if they existed,
library prefixes, and library postfixes, and the order
they are in. This commit should fix that.

For Windows, opencv is thought to be not able to be found
due to an extra version number at the end which isn't
accounted for.